src/requirements.txt
absl-py==2.0.0
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.1
gast==0.5.4
grpcio==1.60.0
h5py==3.10.0
huggingface-hub
idna==3.6
keras
kombu==5.3.4
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy
oauthlib
opt-einsum==3.3.0
packaging==23.2
pandas
prompt-toolkit==3.0.41
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.16.0
sqlparse
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor==2.4.0
tokenizers==0.15.0
tqdm
transformers
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
